requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cssselect==1.2.0
distro==1.9.0
dotenv==0.9.9
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.17.0
google_search_results==2.4.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
lxml==5.3.1
lxml_html_clean==0.4.1
MarkupSafe==3.0.2
newspaper3k==0.2.8
nltk==3.9.1
nodeenv==1.9.1
numpy==2.2.3
openai==1.64.0
pandas==2.2.3
pillow==11.1.0
prisma==0.15.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyupbit==0.2.34
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tinysegmenter==0.3
tldextract==5.1.3
tomlkit==0.13.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websockets==15.0
